REDxEYE got still working at this alpha channel separator, so even if trivial, let it enabled
