+++ texture_tools.py
@@ -15,18 +15,3 @@ def split_alpha(blender_texture):
     alpha_im.pack(as_png=True)
     alpha_view[:] = 1
     blender_texture.pixels = image_data
-
-def split_alpha_menu(self, context):
-    self.layout.operator(IMAGE_MT_AlphaSplit.bl_idname, text='Split alpha')
-
-class IMAGE_MT_AlphaSplit(bpy.types.Operator):
-    """Extracts alpha to new file"""
-    bl_idname = "alpha.split"
-    bl_label = "Split alpha"
-    bl_options = {'UNDO'}
-
-    def execute(self, context):
-        sima = context.space_data
-        ima = sima.image
-        split_alpha(ima)
-        return {'FINISHED'}
